backend/requirements.txt
tensorflow==2.13.0
flask>=2.3.0,<4
flask-cors>=4.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
gunicorn>=21.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0
